+++ order-service/src/main/java/camellia/ecommerce/order_service/dtos/OrderItemDto.java
@@ -7,6 +7,5 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.annotation.JsonInclude.Include;
 
 public record OrderItemDto(@JsonProperty("id") @JsonInclude(Include.NON_EMPTY) UUID publicId,
-                @JsonProperty("product_id") UUID productId, Integer quantity,
-                @JsonProperty("unit_price") Integer unitPrice) {
+        @JsonProperty("product_id") UUID productId, Integer quantity, @JsonProperty("unit_price") Double unitPrice) {
 }
